Move extension to its own file
